requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.1
aiosignal==1.3.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
easyocr==1.7.1
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gdown==5.2.0
huggingface-hub==0.24.5
idna==3.7
imageio==2.34.2
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
lightning-utilities==0.11.6
lmdb==1.5.1
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
ninja==1.11.1.1
nltk==3.8.1
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0
psutil==6.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post5
pyparsing==3.1.2
PySocks==1.7.1
python-bidi==0.6.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
scikit-image==0.24.0
scipy==1.14.0
seaborn==0.13.2
setuptools==72.1.0
shapely==2.0.5
six==1.16.0
soupsieve==2.5
sympy==1.13.1
tifffile==2024.8.10
timm==1.0.8
torch==2.4.0
torchmetrics==1.4.1
torchvision==0.19.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.76
ultralytics-thop==2.0.0
urllib3==2.2.2
yarl==1.9.4